backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality (optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
